1. new const file 2. remove css container 3. removed empty css line 4. removed css icon sizing
