+++ projects/ui-framework/src/lib/form-elements/social/social.component.html
@@ -13,9 +13,7 @@
          [hintMessage]="hintMessage"
          [hideLabelOnFocus]="true"
          (inputEvents)="onInputEvents($event)">
-  <div class="prefix-container"input-prefix>
-    <div class="prefix-text">
-      {{socialSelection.prefix}}
-    </div>
+  <div class="prefix-text" input-prefix>
+    {{socialSelection.prefix}}
   </div>
 </b-input>
