+++ projects/ui-framework/src/lib/form-elements/social/social.const.ts
@@ -1,0 +1,28 @@
+import { Icons } from '../../icons/icons.enum';
+
+export interface SocialInputConfig {
+  icon: Icons;
+  prefix: string;
+}
+
+
+const facebook: SocialInputConfig = {
+  icon: Icons.facebook,
+  prefix: 'www.facebook.com/'
+};
+
+const linkedin: SocialInputConfig = {
+  icon: Icons.linkedin,
+  prefix: 'www.linkedin.com/'
+};
+
+const twitter: SocialInputConfig = {
+  icon: Icons.twitter,
+  prefix: 'www.twitter.com/'
+};
+
+export const socialTypesConfig = {
+  facebook,
+  linkedin,
+  twitter
+};
